refactor(factory): move 'config.factory' to 'factory' package
