--- claf/config/factory/__init__.py
@@ -1,15 +1,0 @@
-
-from claf.config.factory.data_reader import DataReaderFactory
-from claf.config.factory.data_loader import DataLoaderFactory
-from claf.config.factory.model import ModelFactory
-from claf.config.factory.optimizer import OptimizerFactory
-from claf.config.factory.tokens import TokenMakersFactory
-
-
-__all__ = [
-    "DataReaderFactory",
-    "DataLoaderFactory",
-    "ModelFactory",
-    "OptimizerFactory",
-    "TokenMakersFactory",
-]
