+++ claf/factory/base.py
